changing structures

+++ server-automation/stop_docker.sh
@@ -1,7 +1,27 @@
 #!/bin/bash
 
-cd /home/home_admin/immich-app &&
-docker compose down &&
+export services=$(docker ps --filter "label=com.docker.compose.project" -q | xargs docker inspect --format='{{index .Config.Labels "com.docker.compose.project"}}'| sort | uniq)
 
-cd /home/home_admin/homelab/samba &&
-docker compose down
+echo -e "\e[31mFollowing services are running:\e[0m"
+echo $services
+
+for service in $services;
+do
+        echo -e "\e[31mWorking on $service .....\e[0m"
+
+        if [[ "$service" = "samba" ]];
+        then
+                cd ~/homelab/samba &&
+                docker compose down;
+        elif [[ "$service" = "immich" ]];
+        then
+                cd ~/immich-app &&
+                docker compose down;
+        else
+                echo "Did not find $service in configuration";
+        fi
+
+done;
+
+echo -e "\e[31mFinished stopping the following services:\e[0m"
+echo $services
